+++ configuration/src/test/java/org/ehrbase/configuration/config/security/ManagementSecurityConfigurationIT.java
@@ -56,6 +56,7 @@ public class ManagementSecurityConfigurationIT {
                 "management.endpoints.web.exposure.include=health,info,metrics,prometheus,loggers",
                 "management.endpoint.info.enabled=true",
                 "management.endpoint.metrics.enabled=true",
+                "management.endpoint.prometheus.enabled=true",
                 "management.endpoint.loggers.enabled=true",
                 "management.endpoint.health.enabled=true",
                 "management.health.db.enabled=false", // turn off db health checks for tests