+++ rest-openehr/src/main/java/org/ehrbase/rest/openehr/specification/TemplateApiSpecification.java
@@ -21,7 +21,9 @@ import io.swagger.v3.oas.annotations.ExternalDocumentation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.List;
 import org.ehrbase.openehr.sdk.response.dto.TemplateResponseData;
+import org.ehrbase.openehr.sdk.response.dto.ehrscape.TemplateMetaDataDto;
 import org.ehrbase.openehr.sdk.webtemplate.model.WebTemplate;
 import org.springframework.http.ResponseEntity;
 
@@ -34,14 +36,9 @@ public interface TemplateApiSpecification {
             externalDocs =
                     @ExternalDocumentation(
                             url =
-                                    "https://specifications.openehr.org/releases/ITS-REST/latest/definitions.html#definitions-adl-1.4-template-post"))
-    ResponseEntity createTemplateClassic(
-            String openehrVersion,
-            String openehrAuditDetails,
-            String contentType,
-            String accept,
-            String prefer,
-            String template);
+                                    "https://specifications.openehr.org/releases/ITS-REST/latest/definition.html#tag/ADL1.4/operation/definition_template_adl1.4_upload"))
+    ResponseEntity<String> createTemplateClassic(
+            String openehrVersion, String openehrAuditDetails, String prefer, String template);
 
     @Operation(
             tags = "ADL 1.4 TEMPLATE",
@@ -49,8 +46,9 @@ public interface TemplateApiSpecification {
             externalDocs =
                     @ExternalDocumentation(
                             url =
-                                    "https://specifications.openehr.org/releases/ITS-REST/latest/definitions.html#definitions-adl-1.4-template-get"))
-    ResponseEntity getTemplatesClassic(String openehrVersion, String openehrAuditDetails, String accept);
+                                    "https://specifications.openehr.org/releases/ITS-REST/latest/definition.html#tag/ADL1.4/operation/definition_template_adl1.4_list"))
+    ResponseEntity<List<TemplateMetaDataDto>> getTemplatesClassic(
+            String openehrVersion, String openehrAuditDetails, String accept);
 
     @Operation(
             tags = "ADL 1.4 TEMPLATE",
@@ -58,8 +56,8 @@ public interface TemplateApiSpecification {
             externalDocs =
                     @ExternalDocumentation(
                             url =
-                                    "https://specifications.openehr.org/releases/ITS-REST/latest/definitions.html#definitions-adl-1.4-template-get-1"))
-    ResponseEntity getTemplateClassic(
+                                    "https://specifications.openehr.org/releases/ITS-REST/latest/definition.html#tag/ADL1.4/operation/definition_template_adl1.4_get"))
+    ResponseEntity<Object> getTemplateClassic(
             String openehrVersion, String openehrAuditDetails, String accept, String templateId);
 
     @Operation(tags = "ADL 1.4 TEMPLATE", summary = "Get an example composition for the specified template")
@@ -80,7 +78,7 @@ public interface TemplateApiSpecification {
             externalDocs =
                     @ExternalDocumentation(
                             url =
-                                    "https://specifications.openehr.org/releases/ITS-REST/latest/definitions.html#definitions-adl-2-template-post"))
+                                    "https://specifications.openehr.org/releases/ITS-REST/latest/definition.html#tag/ADL2/operation/definition_template_adl2_upload"))
     ResponseEntity<TemplateResponseData> createTemplateNew(
             String openehrVersion,
             String openehrAuditDetails,
@@ -92,6 +90,16 @@ public interface TemplateApiSpecification {
 
     @Operation(
             tags = "ADL 2 TEMPLATE",
+            summary = "List templates",
+            externalDocs =
+                    @ExternalDocumentation(
+                            url =
+                                    "https://specifications.openehr.org/releases/ITS-REST/latest/definition.html#tag/ADL2/operation/definition_template_adl2_get"))
+    ResponseEntity<TemplateResponseData> getTemplatesNew(
+            String openehrVersion, String openehrAuditDetails, String accept);
+
+    @Operation(
+            tags = "ADL 2 TEMPLATE",
             summary = "Get template",
             externalDocs =
                     @ExternalDocumentation(
@@ -100,6 +108,12 @@ public interface TemplateApiSpecification {
     ResponseEntity<TemplateResponseData> getTemplateNew(
             String openehrVersion, String openehrAuditDetails, String accept, String templateId, String versionPattern);
 
-    @Operation(tags = "TEMPLATE", summary = "Get a webtemplate for the specified template")
+    @Operation(
+            tags = "TEMPLATE",
+            summary = "Deprecated since 2.2.0 and marked for removal",
+            description =
+                    "Replaced by [/rest/openehr/v1/definition/template/adl1.4/{template_id}](./index.html?urls.primaryName=1.%20openEHR%20API#/ADL%201.4%20TEMPLATE/getTemplateClassic)",
+            deprecated = true)
+    @Deprecated(since = "2.2.0", forRemoval = true)
     ResponseEntity<WebTemplate> getWebTemplate(String accept, String templateId);
 }