+++ service/src/test/java/org/ehrbase/repository/CompositionRepositoryTest.java
@@ -82,7 +82,7 @@ class CompositionRepositoryTest {
         OffsetDateTime now = OffsetDateTime.of(2022, 3, 21, 23, 45, 10, 123_456_780, ZoneOffset.ofHours(2));
         Mockito.when(timeProvider.getNow()).thenReturn(now);
 
-        DefaultDSLContext context = new DefaultDSLContext(SQLDialect.YUGABYTEDB);
+        DefaultDSLContext context = new DefaultDSLContext(SQLDialect.POSTGRES);
         CompositionRepository repo = new CompositionRepository(context, null, null, null, timeProvider);
 
         VersionDataDbRecord versionData = repo.toRecords(EHR_ID, versionDataObject, CONTRIBUTION_ID, AUDIT_ID);
@@ -95,9 +95,8 @@ class CompositionRepositoryTest {
     private static <R extends org.jooq.Record> String toCsv(DefaultDSLContext context, List<R> dataRecords) {
         Result<R> result = context.newResult(DSL.table(dataRecords.get(0)));
         result.addAll(dataRecords);
-        String dataCsv = result.formatCSV(
+        return result.formatCSV(
                 CSVFormat.DEFAULT.nullString("").emptyString("''").quoteString("'"));
-        return dataCsv;
     }
 
     private static String loadExpectedCsv(String name, boolean version) throws IOException {